+++ src/main/java/com/colonelkatsu/techNotes/controllers/PostController.java
@@ -13,8 +13,9 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
-import com.colonelkatsu.techNotes.models.Account;
-import com.colonelkatsu.techNotes.models.Post;
+
+import com.colonelkatsu.techNotes.entity.Account;
+import com.colonelkatsu.techNotes.entity.Post;
 import com.colonelkatsu.techNotes.services.AccountService;
 import com.colonelkatsu.techNotes.services.PostService;
