+++ src/main/java/com/colonelkatsu/techNotes/entity/Post.java
@@ -1,4 +1,4 @@
-package com.colonelkatsu.techNotes.models;
+package com.colonelkatsu.techNotes.entity;
 
 import java.time.LocalDateTime;
