+++ src/main/java/com/colonelkatsu/techNotes/repositories/AccountRepository.java
@@ -3,7 +3,8 @@ package com.colonelkatsu.techNotes.repositories;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import com.colonelkatsu.techNotes.models.Account;
+
+import com.colonelkatsu.techNotes.entity.Account;
 
 @Repository
 public interface AccountRepository extends JpaRepository<Account, Long> {
