+++ src/main/java/com/colonelkatsu/techNotes/repositories/AuthorityRepository.java
@@ -2,7 +2,8 @@ package com.colonelkatsu.techNotes.repositories;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import com.colonelkatsu.techNotes.models.Authority;
+
+import com.colonelkatsu.techNotes.entity.Authority;
 
 @Repository
 public interface AuthorityRepository extends JpaRepository<Authority, String>  {
