+++ src/main/java/com/colonelkatsu/techNotes/repositories/MessageRepository.java
@@ -3,7 +3,7 @@ package com.colonelkatsu.techNotes.repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import com.colonelkatsu.techNotes.models.Message;
+import com.colonelkatsu.techNotes.entity.Message;
 
 @Repository
 public interface MessageRepository extends JpaRepository<Message, Long> {
