+++ src/main/java/com/colonelkatsu/techNotes/repositories/PostRepository.java
@@ -4,7 +4,8 @@ import java.util.List;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import com.colonelkatsu.techNotes.models.Post;
+
+import com.colonelkatsu.techNotes.entity.Post;
 
 @Repository
 public interface PostRepository extends JpaRepository<Post, Long> {
