+++ src/main/java/com/colonelkatsu/techNotes/services/AccountService.java
@@ -6,7 +6,8 @@ import java.util.Optional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
-import com.colonelkatsu.techNotes.models.Account;
+
+import com.colonelkatsu.techNotes.entity.Account;
 import com.colonelkatsu.techNotes.repositories.AccountRepository;
 
 @Service
