+++ src/main/java/com/colonelkatsu/techNotes/services/MessageService.java
@@ -4,7 +4,8 @@ import java.time.LocalDateTime;
 import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-import com.colonelkatsu.techNotes.models.Message;
+
+import com.colonelkatsu.techNotes.entity.Message;
 import com.colonelkatsu.techNotes.repositories.MessageRepository;
 
 @Service
